backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic-settings
httpx
supabase
google-generativeai
pypdf
